:recycle: Refactors SeekBar name
