+++ emeraldcomponents/src/main/res/layout/widget_score_seek_bar.xml
